migrate channel view customizations to class override mechanism.
